yay more fixes yaaay

+++ src/r_draw.h
@@ -113,7 +113,7 @@ extern boolean skincolor_modified[];
 void R_InitTranslationTables(void);
 UINT8* R_GetTranslationColormap(INT32 skinnum, UINT16 color, UINT8 flags);
 void R_FlushTranslationColormapCache(void);
-UINT8 R_GetColorByName(const char *name);
+UINT16 R_GetColorByName(const char *name);
 UINT16 R_GetSuperColorByName(const char *name);
 
 // Custom player skin translation
